+++ R/Seurat.R
@@ -1,0 +1,12 @@
+#' Seurat
+#'
+#' Example `Seurat` (v4) object.
+#'
+#' @note Updated 2022-06-09.
+#' @usage data(Seurat)
+#' @return `Seurat`.
+#'
+#' @examples
+#' data(Seurat)
+#' print(Seurat)
+"Seurat"