+++ R/SeuratMarkersPerCluster.R
@@ -1,0 +1,18 @@
+## NOTE Hitting some roxygen weirdness with this object.
+##
+## Warning: The ncol() method for CompressedSplitDataFrameList objects is
+## deprecated. Please use ncols() on these objects instead.
+##
+## - https://github.com/r-lib/roxygen2/issues/595
+## - https://github.com/ComputationalProteomicsUnit/maker/issues/28
+
+#' Example `SeuratMarkersPerCluster` object
+#'
+#' @note Updated 2022-06-09.
+#' @usage data(SeuratMarkersPerCluster)
+#' @return `SeuratMarkersPerCluster`.
+#'
+#' @examples
+#' data(SeuratMarkersPerCluster)
+#' print(SeuratMarkersPerCluster)
+"SeuratMarkersPerCluster"